+++ Classes/Backend/Form/FormDataProvider/PageLanguageOverlayFormDataProvider.php
@@ -1,0 +1,16 @@
+<?php
+namespace T3v\T3vBase\Backend\Form\FormDataProvider;
+
+use T3v\T3vBase\Backend\Form\FormDataProvider\PageFormDataProvider;
+
+/**
+ * The page language overlay form data provider class.
+ *
+ * @package T3v\T3vBase\Backend\Form\FormDataProvider
+ */
+class PageLanguageOverlayFormDataProvider extends PageFormDataProvider {
+  /**
+   * The table name of the form data provider.
+   */
+  const TABLE_NAME = 'pages_language_overlay';
+}