+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Added `PageFormDataProvider` and `PageLanguageOverlayFormDataProvider`
 * Added common files
 
 [Semantic Versioning]: http://semver.org "Semantic Versioning"